feat: Move Query tests to Android tests
MLIBZ-1868
