+++ android-lib/build.gradle
@@ -47,7 +47,7 @@ android {
                 events "failed"
                 exceptionFormat "full"
             }
-            setIgnoreFailures(true)
+//            setIgnoreFailures(true)
         }
     }
 }
