+++ android-lib/src/androidTest/assets/kinvey.properties
@@ -1,6 +1,6 @@
 # Required settings
-app.key=<YOUR_APP_KEY_HERE>
-app.secret=<YOUR_APP_SECRET_HERE>
+app.key=kid_BkssrV2eW
+app.secret=dadb9b61c32c48bea813ea9c3e116be7
 
 # Optional settings, only required for push
