+++ android-lib/src/androidTest/java/com/kinvey/androidTest/cache/QueryTest.java
@@ -38,13 +38,12 @@ import static org.mockito.Mockito.verify;
 @SmallTest
 public class QueryTest {
 
-    RealmQuery<DynamicRealmObject> query;
+    private RealmQuery<DynamicRealmObject> query;
+    private static final String TEST_FIELD = "field";
 
-    ICacheManager cacheManager;
     @Before
     public void setup(){
         Context mMockContext = new RenamingDelegatingContext(InstrumentationRegistry.getInstrumentation().getTargetContext(), "test_");
-        cacheManager = new Client.Builder(mMockContext).build().getCacheManager();
         Realm.init(mMockContext);
         DynamicRealm realm = DynamicRealm.getInstance(new RealmConfiguration.Builder().build());
 
@@ -100,6 +99,238 @@ public class QueryTest {
         verify(query, times(1)).or();
         verify(query, times(2)).endGroup();
         verify(query, times(1)).not();
+    }
+
+    @Test
+    public void testGreaterThanIntClause() {
+        Query q = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        q.greaterThan(TEST_FIELD, 1);
+        QueryHelper.prepareRealmQuery(query, q.getQueryFilterMap());
+        verify(query, times(1)).greaterThan(TEST_FIELD, 1);
+    }
+
+    @Test
+    public void testGreaterThanLongClause() {
+        Query q = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        q.greaterThan(TEST_FIELD, 1L);
+        QueryHelper.prepareRealmQuery(query, q.getQueryFilterMap());
+        verify(query, times(1)).greaterThan(TEST_FIELD, 1L);
+    }
+
+    @Test
+    public void testGreaterThanDoubleClause() {
+        Query q = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        q.greaterThan(TEST_FIELD, 1.0d);
+        QueryHelper.prepareRealmQuery(query, q.getQueryFilterMap());
+        verify(query, times(1)).greaterThan(TEST_FIELD, 1.0d);
+    }
+
+    @Test
+    public void testGreaterThanFloatClause() {
+        Query q = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        q.greaterThan(TEST_FIELD, 1.0f);
+        QueryHelper.prepareRealmQuery(query, q.getQueryFilterMap());
+        verify(query, times(1)).greaterThan(TEST_FIELD, 1.0f);
+    }
+
+    @Test
+    public void testGreaterThanOrEqualToIntClause() {
+        Query q = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        q.greaterThanEqualTo(TEST_FIELD, 1);
+        QueryHelper.prepareRealmQuery(query, q.getQueryFilterMap());
+        verify(query, times(1)).greaterThanOrEqualTo(TEST_FIELD, 1);
+    }
+
+    @Test
+    public void testGreaterThanOrEqualToLongClause() {
+        Query q = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        q.greaterThanEqualTo(TEST_FIELD, 1L);
+        QueryHelper.prepareRealmQuery(query, q.getQueryFilterMap());
+        verify(query, times(1)).greaterThanOrEqualTo(TEST_FIELD, 1L);
+    }
+
+    @Test
+    public void testGreaterThanOrEqualToDoubleClause() {
+        Query q = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        q.greaterThanEqualTo(TEST_FIELD, 1.0d);
+        QueryHelper.prepareRealmQuery(query, q.getQueryFilterMap());
+        verify(query, times(1)).greaterThanOrEqualTo(TEST_FIELD, 1.0d);
+    }
+
+    @Test
+    public void testGreaterThanOrEqualToFloatClause() {
+        Query q = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        q.greaterThanEqualTo(TEST_FIELD, 1.0f);
+        QueryHelper.prepareRealmQuery(query, q.getQueryFilterMap());
+        verify(query, times(1)).greaterThanOrEqualTo(TEST_FIELD, 1.0f);
+    }
+
+
+    @Test
+    public void testLessThanIntClause() {
+        Query q = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        q.lessThan(TEST_FIELD, 1);
+        QueryHelper.prepareRealmQuery(query, q.getQueryFilterMap());
+        verify(query, times(1)).lessThan(TEST_FIELD, 1);
+    }
+
+    @Test
+    public void testLessThanLongClause() {
+        Query q = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        q.lessThan(TEST_FIELD, 1L);
+        QueryHelper.prepareRealmQuery(query, q.getQueryFilterMap());
+        verify(query, times(1)).lessThan(TEST_FIELD, 1L);
+    }
+
+    @Test
+    public void testLessThanDoubleClause() {
+        Query q = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        q.lessThan(TEST_FIELD, 1.0d);
+        QueryHelper.prepareRealmQuery(query, q.getQueryFilterMap());
+        verify(query, times(1)).lessThan(TEST_FIELD, 1.0d);
+    }
+
+    @Test
+    public void testLessThanFloatClause() {
+        Query q = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        q.lessThan(TEST_FIELD, 1.0f);
+        QueryHelper.prepareRealmQuery(query, q.getQueryFilterMap());
+        verify(query, times(1)).lessThan(TEST_FIELD, 1.0f);
+    }
+
+    @Test
+    public void testLessThanOrEqualToIntClause() {
+        Query q = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        q.lessThanEqualTo(TEST_FIELD, 1);
+        QueryHelper.prepareRealmQuery(query, q.getQueryFilterMap());
+        verify(query, times(1)).lessThanOrEqualTo(TEST_FIELD, 1);
+    }
+
+    @Test
+    public void testLessThanOrEqualToLongClause() {
+        Query q = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        q.lessThanEqualTo(TEST_FIELD, 1L);
+        QueryHelper.prepareRealmQuery(query, q.getQueryFilterMap());
+        verify(query, times(1)).lessThanOrEqualTo(TEST_FIELD, 1L);
+    }
+
+    @Test
+    public void testLessThanOrEqualToDoubleClause() {
+        Query q = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        q.lessThanEqualTo(TEST_FIELD, 1.0d);
+        QueryHelper.prepareRealmQuery(query, q.getQueryFilterMap());
+        verify(query, times(1)).lessThanOrEqualTo(TEST_FIELD, 1.0d);
+    }
+
+    @Test
+    public void testLessThanOrEqualToFloatClause() {
+        Query q = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        q.lessThanEqualTo(TEST_FIELD, 1.0f);
+        QueryHelper.prepareRealmQuery(query, q.getQueryFilterMap());
+        verify(query, times(1)).lessThanOrEqualTo(TEST_FIELD, 1.0f);
+    }
+
+    @Test
+    public void testNotEqualBooleanClause() {
+        Query q = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        q.notEqual(TEST_FIELD, true);
+        QueryHelper.prepareRealmQuery(query, q.getQueryFilterMap());
+        verify(query, times(1)).notEqualTo(TEST_FIELD, true);
+    }
+
+    @Test
+    public void testNotEqualByteClause() {
+        Query q = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        Byte aByte = 1;
+        q.notEqual(TEST_FIELD, aByte);
+        QueryHelper.prepareRealmQuery(query, q.getQueryFilterMap());
+        verify(query, times(1)).notEqualTo(TEST_FIELD, aByte);
+    }
+
+    @Test
+    public void testNotEqualByteArrayClause() {
+        Query q = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        byte[] bytes = {1};
+        q.notEqual(TEST_FIELD, bytes);
+        QueryHelper.prepareRealmQuery(query, q.getQueryFilterMap());
+        verify(query, times(1)).notEqualTo(TEST_FIELD, bytes);
+    }
+
+    @Test
+    public void testNotEqualShortClause() {
+        Query q = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        Short aShort = 12345;
+        q.notEqual(TEST_FIELD, aShort);
+        QueryHelper.prepareRealmQuery(query, q.getQueryFilterMap());
+        verify(query, times(1)).notEqualTo(TEST_FIELD, aShort);
+    }
+
+    @Test
+    public void testNotEqualIntClause() {
+        Query q = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        q.notEqual(TEST_FIELD, 1);
+        QueryHelper.prepareRealmQuery(query, q.getQueryFilterMap());
+        verify(query, times(1)).notEqualTo(TEST_FIELD, 1);
+    }
+
+    @Test
+    public void testNotEqualLongClause() {
+        Query q = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        q.notEqual(TEST_FIELD, 1L);
+        QueryHelper.prepareRealmQuery(query, q.getQueryFilterMap());
+        verify(query, times(1)).notEqualTo(TEST_FIELD, 1L);
+    }
+
+    @Test
+    public void testNotEqualDoubleClause() {
+        Query q = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        q.notEqual(TEST_FIELD, 1d);
+        QueryHelper.prepareRealmQuery(query, q.getQueryFilterMap());
+        verify(query, times(1)).notEqualTo(TEST_FIELD, 1d);
+    }
+
+    @Test
+    public void testNotEqualFloatClause() {
+        Query q = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        q.notEqual(TEST_FIELD, 1f);
+        QueryHelper.prepareRealmQuery(query, q.getQueryFilterMap());
+        verify(query, times(1)).notEqualTo(TEST_FIELD, 1f);
+    }
+
+    @Test
+    public void testNotEquqlStringClause() {
+        Query q = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        q.notEqual(TEST_FIELD, TEST_FIELD);
+        QueryHelper.prepareRealmQuery(query, q.getQueryFilterMap());
+        verify(query, times(1)).notEqualTo(TEST_FIELD, TEST_FIELD);
+    }
+
+    @Test
+    public void testOrClause() {
+        Query q = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        Query q2 = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        q.in(TEST_FIELD, new String[]{"1", "2"});
+        q2.in(TEST_FIELD, new String[]{"1", "2"});
+        q.or(q2);
+        QueryHelper.prepareRealmQuery(query, q.getQueryFilterMap());
+        verify(query, times(5)).beginGroup();
+        verify(query, times(4)).equalTo(eq(TEST_FIELD), anyString());
+        verify(query, times(3)).or();
+        verify(query, times(5)).endGroup();
+    }
+
+    @Test
+    public void testAndClause() {
+        Query q = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        Query q2 = new Query(new MongoQueryFilter.MongoQueryFilterBuilder());
+        q.notEqual(TEST_FIELD, 1);
+        q2.notEqual(TEST_FIELD, 2);
+        q.and(q2);
+        QueryHelper.prepareRealmQuery(query, q.getQueryFilterMap());
+        verify(query, times(1)).notEqualTo(TEST_FIELD, 1);
+        verify(query, times(1)).notEqualTo(TEST_FIELD, 2);
+        verify(query, times(3)).beginGroup();
+        verify(query, times(3)).endGroup();
     }
 
 }
