--- android-lib/src/test/java/com/kinvey/android/util/PersonRepository.java
@@ -1,6 +1,0 @@
-package com.kinvey.android.util;
-
-
-public interface PersonRepository {
-    void createPerson(String name);
-}
